feature/segregated-instrumentation
- isolate flutter drive instrumentation from test app
- move integration tests to their own file
